+++ SW1_ISO9126_FUZZY/Vistas/EvaluacionSWPage.xaml
@@ -1,0 +1,54 @@
+﻿<Page x:Class="SW1_ISO9126_FUZZY.Vistas.EvaluacionSWPage"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:SW1_ISO9126_FUZZY.Vistas"
+      xmlns:controls="http://metro.mahapps.com/winfx/xaml/controls"
+      mc:Ignorable="d" 
+      d:DesignHeight="300" d:DesignWidth="300"
+      Title="EvaluacionSWPage" Width="800" Height="600" Background="#FF80BA45">
+
+    <Grid>
+        <Grid Background="White">
+            <GroupBox Header="Características Internas" HorizontalAlignment="Left" Height="251" Margin="10,10,0,0" VerticalAlignment="Top" Width="780" OpacityMask="Black" Background="#FF33A133" BorderBrush="#FF33A133">
+                <Grid>
+                    <controls:Tile x:Name="btnFuncInterna" Title="Funcionabilidad" Margin="74,28,554,44" Click="btnFuncInterna_Click" Background="#FFFF3A00">
+                        <Image Source="/SW1_ISO9126_FUZZY;component/Imagenes/Agreement-01.png" Height="130" Width="130"></Image>
+                    </controls:Tile>
+                    <controls:Tile x:Name="btnMantInterna" Title="Mantenibilidad" Margin="328,28,300,44" Click="btnMantInterna_Click" Background="#FFFF3A00">
+                        <Image Source="/SW1_ISO9126_FUZZY;component/Imagenes/Toolbox.png" Height="130" Width="130"></Image>
+                    </controls:Tile>
+                    <controls:Tile x:Name="btnUsabInterna" Title="Usabilidad" Margin="575,28,53,44" TiltFactor="6" Click="btnUsabInterna_Click" Background="#FFFF3A00">
+                        <Image Source="/SW1_ISO9126_FUZZY;component/Imagenes/Hand-Sign-06.png" Height="130" Width="130"></Image>
+                    </controls:Tile>
+                    <Label x:Name="cantFuncInterna" Content="10/19" Margin="118,168,610,17" />
+                    <Label x:Name="cantUsabInterna" Content="10/19" Margin="378,168,350,17" />
+                    <Label x:Name="cantMantInterna" Content="10/19" Margin="625,168,103,17" />
+                </Grid>
+            </GroupBox>
+            <GroupBox Header="Características Externas" HorizontalAlignment="Left" Height="251" Margin="10,266,0,0" VerticalAlignment="Top" Width="780" BorderBrush="#FF33A133" Background="#FF33A133">
+                <Grid>
+                    <controls:Tile x:Name="btnFuncExterna" Title="Funcionabilidad" Margin="74,28,554,44" Click="btnFuncExterna_Click" Background="#FFFF3A00">
+                        <Image Source="/SW1_ISO9126_FUZZY;component/Imagenes/Branch-Engineering.png" Height="130" Width="130"></Image>
+                    </controls:Tile>
+                    <controls:Tile x:Name="btnMantExterna" Title="Mantenibilidad" Margin="325,28,303,44" Click="btnMantExterna_Click" Background="#FFFF3A00">
+                        <Image Source="/SW1_ISO9126_FUZZY;component/Imagenes/Tools-01.png" Height="130" Width="130"></Image>
+                    </controls:Tile>
+                    <controls:Tile x:Name="btnUsabExterna" Title="Usabilidad" Margin="578,28,50,44" Click="btnUsabExterna_Click" Background="#FFFF3A00">
+                        <Image Source="/SW1_ISO9126_FUZZY;component/Imagenes/Touch.png" Height="130" Width="130"></Image>
+                    </controls:Tile>
+                    <Label x:Name="cantFuncExterna" Content="10/19" Margin="118,168,610,17" />
+                    <Label x:Name="cantUsabExterna" Content="10/19" Margin="372,168,356,17" />
+                    <Label x:Name="cantMantExterna" Content="10/19" Margin="630,173,98,12" />
+                </Grid>
+            </GroupBox>
+            <Grid HorizontalAlignment="Left" Height="29" Margin="451,522,0,0" VerticalAlignment="Top" Width="339">
+                <Button Content="Siguiente" HorizontalAlignment="Left" Margin="679,486,0,0" VerticalAlignment="Top" Width="75" RenderTransformOrigin="0.635,-0.217"/>
+                <Button Content="Volver" HorizontalAlignment="Left" Margin="565,486,0,0" VerticalAlignment="Top" Width="75"/>
+
+            </Grid>
+
+        </Grid>
+    </Grid>
+</Page>
