+++ SW1_ISO9126_FUZZY/Vistas/RegistroSWPage.xaml.cs
@@ -1,0 +1,19 @@
+﻿using System.Windows;
+using System.Windows.Controls;
+using System;
+using SW1_ISO9126_FUZZY.Archivos;
+using Xceed.Wpf.Toolkit;
+
+
+namespace SW1_ISO9126_FUZZY.Vistas {
+    /// <summary>
+    /// Lógica de interacción para MainPage.xaml
+    /// </summary>
+    public partial class MainPage : Page {
+        public MainPage() {
+            InitializeComponent();
+        }
+
+        
+    }
+}
